BF: Treat git-mv'ed annexed files like untracked content
git-mv stages the move, and our handling of such a change causes
a potential fix-up of an annex symlink to happen outside and after
our change analysis. This change unstages such paths and
re-categorizes such a modification as an untracked addition. This
essentially removes the special case and fixes gh-4967